+++ workflows/index.ts
@@ -22,6 +22,7 @@ export function concreteWorkflowToMeta(
 export * from "./src/xstate.ts";
 export * from "./src/copy-template-machine/copy-template-machine.ts";
 export * from "./src/copy-template-machine/types.ts";
+export * from "./src/update-template-machine/update-template-machine.ts";
 export * from "./src/test-runner.ts";
 export * from "./src/chain.ts";
 export * from "./src/npm.ts";
