+++ workflows/src/types.ts
@@ -1,4 +1,5 @@
 import type { Snapshot } from "xstate";
+import type { WorkflowContext } from "./xstate.ts";
 
 export type Result<C extends Record<string, any>> = {
   data?: C;
@@ -29,4 +30,11 @@ export interface WorkflowBlob {
   workflowName: string;
   internalState?: WorkflowBlobInternalState;
   snapshotState?: Snapshot<any>;
+}
+
+export interface TemplateWorkflowContext extends WorkflowContext {
+  name: string;
+  pascalName: string;
+  targetDir: string;
+  sourceDir: string;
 }
